loadgen/requirements.txt
# ./loadgen/requirements.txt
psycopg2-binary==2.9.9
